Inicio da transformação da lógica de negócios do aplicativo
A lógica de negócios vai passar por uma refatoração para ficar mais elegante e seguir boas práticas e esse é o primeiro commit desse processo.
