+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"F:\\GitHub\\virus-tracker-tm\\src\\Components\\NavBar\\NavBar.tsx":"1","F:\\GitHub\\virus-tracker-tm\\src\\Components\\TextBox\\TextBox.tsx":"2","F:\\GitHub\\virus-tracker-tm\\src\\index.tsx":"3","F:\\GitHub\\virus-tracker-tm\\src\\Components\\Searchbar\\Searchbar.tsx":"4","F:\\GitHub\\virus-tracker-tm\\src\\Components\\App\\App.tsx":"5","F:\\GitHub\\virus-tracker-tm\\src\\components\\searchBar\\Searchbar.tsx":"6","F:\\GitHub\\virus-tracker-tm\\src\\api\\covidAPI.ts":"7","F:\\GitHub\\virus-tracker-tm\\src\\utils\\NAOCOMITAISSOAQUI.ts":"8","F:\\GitHub\\virus-tracker-tm\\src\\components\\app\\App.tsx":"9","F:\\GitHub\\virus-tracker-tm\\src\\components\\TextCity\\TextCity.tsx":"10","F:\\GitHub\\virus-tracker-tm\\src\\constants.ts":"11","F:\\GitHub\\virus-tracker-tm\\src\\components\\navBar\\NavBar.tsx":"12","F:\\GitHub\\virus-tracker-tm\\src\\components\\textBox\\TextBox.tsx":"13","F:\\GitHub\\virus-tracker-tm\\src\\components\\ItemList\\ItemList.tsx":"14","F:\\GitHub\\virus-tracker-tm\\src\\components\\MenuIcon\\MenuIcon.tsx":"15"},{"size":812,"mtime":1611621814771,"results":"16","hashOfConfig":"17"},{"size":866,"mtime":1611343738141,"results":"18","hashOfConfig":"17"},{"size":378,"mtime":1611866768977,"results":"19","hashOfConfig":"17"},{"size":2177,"mtime":1611634009058,"results":"20","hashOfConfig":"17"},{"size":1695,"mtime":1611636139607,"results":"21","hashOfConfig":"17"},{"size":2434,"mtime":1611870215847,"results":"22","hashOfConfig":"17"},{"size":386,"mtime":1611777543534,"results":"23","hashOfConfig":"17"},{"size":71,"mtime":1611636744000,"results":"24","hashOfConfig":"17"},{"size":1924,"mtime":1611870535848,"results":"25","hashOfConfig":"17"},{"size":756,"mtime":1611870862158,"results":"26","hashOfConfig":"17"},{"size":121,"mtime":1611867072916,"results":"27","hashOfConfig":"17"},{"size":2873,"mtime":1611955526270,"results":"28","hashOfConfig":"17"},{"size":911,"mtime":1611870902495,"results":"29","hashOfConfig":"17"},{"size":837,"mtime":1611955702546,"results":"30","hashOfConfig":"17"},{"size":418,"mtime":1611953966866,"results":"31","hashOfConfig":"17"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1pvhbp5",{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"56","messages":"57","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"58","messages":"59","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"60","messages":"61","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"F:\\GitHub\\virus-tracker-tm\\src\\Components\\NavBar\\NavBar.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\Components\\TextBox\\TextBox.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\index.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\Components\\Searchbar\\Searchbar.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\Components\\App\\App.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\searchBar\\Searchbar.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\api\\covidAPI.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\utils\\NAOCOMITAISSOAQUI.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\app\\App.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\TextCity\\TextCity.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\constants.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\navBar\\NavBar.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\textBox\\TextBox.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\ItemList\\ItemList.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\MenuIcon\\MenuIcon.tsx",[]]
+[{"F:\\GitHub\\virus-tracker-tm\\src\\index.tsx":"1","F:\\GitHub\\virus-tracker-tm\\src\\api\\covidAPI.ts":"2","F:\\GitHub\\virus-tracker-tm\\src\\utils\\NAOCOMITAISSOAQUI.ts":"3","F:\\GitHub\\virus-tracker-tm\\src\\constants.ts":"4","F:\\GitHub\\virus-tracker-tm\\src\\components\\MenuIcon.tsx":"5","F:\\GitHub\\virus-tracker-tm\\src\\components\\App.tsx":"6","F:\\GitHub\\virus-tracker-tm\\src\\components\\NavBar.tsx":"7","F:\\GitHub\\virus-tracker-tm\\src\\components\\TextBox.tsx":"8","F:\\GitHub\\virus-tracker-tm\\src\\components\\ItemList.tsx":"9","F:\\GitHub\\virus-tracker-tm\\src\\components\\Searchbar.tsx":"10"},{"size":378,"mtime":1611866768977,"results":"11","hashOfConfig":"12"},{"size":418,"mtime":1612716149202,"results":"13","hashOfConfig":"12"},{"size":71,"mtime":1611636744000,"results":"14","hashOfConfig":"12"},{"size":121,"mtime":1611867072916,"results":"15","hashOfConfig":"12"},{"size":415,"mtime":1612715934491,"results":"16","hashOfConfig":"12"},{"size":1782,"mtime":1612718137801,"results":"17","hashOfConfig":"12"},{"size":2922,"mtime":1612716004230,"results":"18","hashOfConfig":"12"},{"size":810,"mtime":1612715327158,"results":"19","hashOfConfig":"12"},{"size":894,"mtime":1612715779592,"results":"20","hashOfConfig":"12"},{"size":2293,"mtime":1612716386169,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1pvhbp5",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"F:\\GitHub\\virus-tracker-tm\\src\\index.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\api\\covidAPI.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\utils\\NAOCOMITAISSOAQUI.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\constants.ts",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\MenuIcon.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\App.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\NavBar.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\TextBox.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\ItemList.tsx",[],"F:\\GitHub\\virus-tracker-tm\\src\\components\\Searchbar.tsx",[]]
