+++ src/index.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./components/app/App";
+import App from "./components/App";
 
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 ReactDOM.render(
-  <BrowserRouter>
-    <Switch>
-      <Route path="/" exact={true} component={App} />
-    </Switch>
-  </BrowserRouter>,
-  document.getElementById("root")
+	<BrowserRouter>
+		<Switch>
+			<Route path="/" exact={true} component={App} />
+		</Switch>
+	</BrowserRouter>,
+	document.getElementById("root")
 );
